src/main/java/com/example/userApp/controller/UserController.java
package com.example.userApp.controller;

import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.example.userApp.model.Document;
import com.example.userApp.model.User;
import com.example.userApp.service.UserService;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;

@RestController
@RequestMapping("/api/user")
public class UserController {
    @Autowired
    UserService userService;

    @GetMapping()
    public ResponseEntity<?> getUserByDocument(@RequestParam Integer documentId,
            @RequestParam String documentType) {
        try {
            Document documentTypeTyped = Document.valueOf(documentType);
            User user = userService.findUserByDocument(documentId, documentTypeTyped);

            // Si no se encuentra el usuario, se devuelve un 404 pero debería retornar un
            // 204(no content)
            if (user == null) {
                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User not found");
            }
            return ResponseEntity.ok(user);
        } catch (IllegalArgumentException e) {
            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid document type: " + documentType);
        } catch (Exception e) {
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .body("An unexpected error occurred: " + e.getMessage());
        }
    }

}
